Implement profile photo handling: add image download, storage, and display functionality; update UI components to show profile photos

+++ app/src/commonMain/kotlin/com/nkwabyte/medilert/util/Sha1.kt
@@ -40,7 +40,7 @@ private fun sha1Bytes(message: ByteArray): ByteArray {
             when {
                 i < 20 -> { f = (b and c) or (b.inv() and d); k = 0x5A827999 }
                 i < 40 -> { f = b xor c xor d;                k = 0x6ED9EBA1 }
-                i < 60 -> { f = (b and c) or (b and d) or (c and d); k = 0x8F1BBCDC }
+                i < 60 -> { f = (b and c) or (b and d) or (c and d); k = 0x8F1BBCDC.toInt() }
                 else   -> { f = b xor c xor d;                k = 0xCA62C1D6.toInt() }
             }
             val temp = ((a shl 5) or (a ushr 27)) + f + e + k + w[i]
